@atgeirr 2016-02-01
First, let me say that I think this looks good. We'll merge this eventually. Some comments and questions:
 - This PR does not (on purpose I assume) account for temperature derivatives. Adding it should not be hard, but would increase the size of the Evaluation objects by one, and therefore probably reduce performance somewhat. Is there a way to avoid this when we don't have to? We can check the input AutoDiffBlock for temperature if it is constant to see if this is needed (using `.derivatives().empty()`, we should add a clearer `isConstant()` for that I guess).
 - There seems to be a lot of almost-repetition in setting up the local AD objects and fluid states. Can this be reduced (given the constraint that the inputs are still AutoDiffBlock objects)?
 - Do you think the tag system for the Evaluation class brings genuine security from errors? To me it seems about as likely that someone would mess up the tags as well, if they are making the error they are designed to prevent.
 - Do we have now (in opm-material) a fluid system that supports the black-oil things + temperature dependence?

Back to the PR: I'll merge this when @totto82 confirms that the results are good and matching. I appreciate that you have made the effort to achieve this, hopefully this check does not reveal any delaying errors.