@andlaus 2016-02-01
> This PR does not (on purpose I assume) account for temperature derivatives. Adding it should not be hard, but would increase the size of the Evaluation objects by one, and therefore probably reduce performance somewhat. Is there a way to avoid this when we don't have to? We can check the input AutoDiffBlock for temperature if it is constant to see if this is needed (using .derivatives().empty(), we should add a clearer isConstant() for that I guess).

that's true: With this PR I wanted it to recreate _exactly_ what the current master provides. For implementing proper temperature support, I think a bit of copy-and-pasting akin

```c++
if (needTemperature) {
    // include temperature in the LAD variables and apply the chain rule
}
else {
    // roll with what's currently done
}
```

would be okay? the `needTemperature` check could be done as you propose. (note that this avoids regressing performance of isothermal simulations because of of the thermal code.)

> There seems to be a lot of almost-repetition in setting up the local AD objects and fluid states. Can this be reduced (given the constraint that the inputs are still AutoDiffBlock objects)?

I'm not really sure if that's possible: It would certainly be possible to write a dynamically sized `Evaluation` which is based on `ADB`, but performance-wise I'm not so sure if this improves things because the memory access pattern will be more random (thus prefetching might work less well), dynamically-sized objects prevent all kind of compiler optimizations (like effective AVX code) and they will need to use the heap more. On the other hand, this would allow to compute the "primary variable derivatives" directly, which might compensate this.

> Do you think the tag system for the Evaluation class brings genuine security from errors? To me it seems about as likely that someone would mess up the tags as well, if they are making the error they are designed to prevent.

I'm not really sure what you mean here: The way I intended it, the 'tag' classes are just an development aid for the programmer because the name of these "tags" appears in the compiler messages. So: No it is not bullet-proof but it was not indented to be that way.

possibly it should be removed, I'm open to that.

> Do we have now (in opm-material) a fluid system that supports the black-oil things + temperature dependence?

yes. fluid-wise the only thing missing for a proper energy conservation equation is enthalpy. (and also the heat capacity and conductivity of the rock, but that's not part of the fluid system.)