@totto82 2016-02-11
I have tested this PR on model 2 and it reproduces the masters results within numerical noise.