@atgeirr 2016-02-11
> it reproduces the masters results within numerical noise

Good news! I'll merge this later today.