@andlaus 2016-02-11
I added a patch to fix a few some typos which I've found in `BlackoilPropsAdFromDeck`. most of them seem to be copy and paste errors stemming from the original code and all of them definitively do not change the result in any way.