@GitPaean 2016-02-12
This is a nice improvement, which gives a 5%-10% performance improvement for Norne with current default parameters.