@GitPaean 2016-02-18
Is this related to this PR?
`sim_poly_fi2p_comp_ad` can not run two-phase simulation now?

```
~/OPM/debug/opm-autodiff-build/bin/sim_poly_fi2p_comp_ad run.param 

================    Test program for fully implicit three-phase black-oil flow     ===============

---------------    Reading parameters     ---------------
deck_filename found at /, value is OWP_OPM.DATA
output not found. Using default value 'true'.
output_dir found at /, value is 'opm-simulation'.
pvt_tab_size found at /, value is '-1'.
threephase_model found at /, value is 'gwseg'.
Program threw an exception: [/home/kaib/OPM/debug/opm-material/opm/material/fluidsystems/blackoilpvt/GasPvtMultiplexer.hpp:128] Not implemented: Gas PVT of this deck!
terminate called after throwing an instance of 'std::logic_error'
  what():  [/home/kaib/OPM/debug/opm-material/opm/material/fluidsystems/blackoilpvt/GasPvtMultiplexer.hpp:128] Not implemented: Gas PVT of this deck!
```