@andlaus 2016-02-19
> Is this related to this PR?

I think so. I suppose gas is disabled in that deck? (If so, we just need to check if a given phase is enabled before we try to initialize its PVT object.)