@GitPaean 2016-02-19
> I suppose gas is disabled in that deck? 

Yes. It is a two-phase case. No gas is specified. Could you fix it to make the two-phase example running again?