@andlaus 2016-02-19
> Yes. It is a two-phase case. No gas is specified. Could you fix it to make the two-phase example running again?

I guess I can make it run. which keywords are used to specify gas in that deck? (I'd look it up myselfm but the deck is probably not shipped with the opm-autodiff sources, right?)