@GitPaean 2016-02-19
I think it is `WATER` for water phase, `OIL` for oil, `GAS` for gas. For oil-water simulation, basically no `GAS` in the deck means no gas phase is involved. 

Basically, we can not use the old style `phase_usage_.phase_used[Liquid]` for the conditions now?

The deck is not in the opm-autodiff. Maybe we should add a test for different of phase combination to avoid some breaking for the same reason happen in the future.