@andlaus 2016-02-19
does OPM/opm-material/127 fix the issue? if yes, please press Green there.